[Console] Fix exit status on uncaught exception with negative code
As described in #45850, if an application threw an exception with
the `code` property set to a negative number this could in some
cases cause the process to appear to exit 'successfully' with a
zero exit status.

Exiting with any negative number produces potentially unexpected
results - the reported exit status will not appear to match the
value that was set. This is due to the binary handling /
truncation of exit codes.

This may theoretically break BC for applications that were
intentionally throwing exceptions with negative codes and
performing some action based on that status. However, given they
would have had to implement an algorithm to map e.g. `-10` in
PHP to `246` as the actual exit status, it seems unlikely this
is a common usage. It is certainly outside the defined behaviour
of POSIX exit codes.

Therefore I believe it is essentially safe to assume that exceptions
with negative codes are e.g. being thrown by lower-level components,
and are not intended to set a shell exit status.

Coalescing all negative numbers to 1 matches the existing behaviour
with other 'invalid' exception codes e.g. empty / zero / non-numeric.

This therefore feels the most robust fix and eliminates any potential
for confusion.

Fixes #45850